tensorflow/requirements.txt
tensorflow 
matplotlib
pillow